pyproject.toml
[tool.poetry]
name = "kym"
version = "0.1.3"
description = "Know Your Model toolkit, to shed light on your blackbox models"
authors = ["soroush.moazed <soroush.moazed@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
omegaconf = "^2.3.0"
pandas = "^1.5.2"
click = "^8.1.3"
tensorflow = ">=2.10.0"
dash = "^2.7.1"
pillow = "^9.3.0"
aimedic = {version = "0.1.2", source = "internal"}
memory-profiler = "^0.61.0"

[[tool.poetry.source]]
name = "internal"
url = "https://pypi.aimedic.tech"
default = false
secondary = false

[tool.poetry.scripts]
kym = "kym.cli:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pre-commit = "^2.21.0"
black = "^22.12.0"
flake8 = "^6.0.0"
mkdocs = "^1.4.2"
mkdocs-material = "^9.0.3"
mkdocstrings = "^0.19.1"
mkdocstrings-python = "^0.8.3"
flake8-docstrings = "^1.6.0"
isort = "^5.11.4"
coverage = "^7.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
check = false        # false: Check code & format it - true: Just check code

[tool.pytest.ini_options]
#addopts = "-v --tb=line -rfEPpxX --cov-report=term-missing --cov=tests --cov-config .coveragerc --disable-warnings"
norecursedirs = "*.egg .eggs dist build docs .tox"
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
